# docs/swagger/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Argo-cd like GitDevOps Continuous Development platform for docker swarm.",
        "title": "Meltcd API",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "https://github.com/kunalsin9h/meltcd/blob/main/LICENSE"
        },
        "version": "0.6"
    },
    "host": "localhost:11771",
    "basePath": "/api",
    "paths": {
...
        "/repo": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": [],
                        "cookies": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Repo"
                ],
                "summary": "Get a list all repositories",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/repo.ListData"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": [],
                        "cookies": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Repo"
                ],
                "summary": "Update a repository",
                "parameters": [
                    {
                        "description": "Repository details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/repo.PrivateRepoDetails"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/app.GlobalResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/app.GlobalResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.GlobalResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": [],
                        "cookies": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Repo"
                ],
                "summary": "Add a new repository",
                "parameters": [
                    {
                        "description": "Repository details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/repo.PrivateRepoDetails"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/app.GlobalResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/app.GlobalResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.GlobalResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": [],
                        "cookies": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Repo"
                ],
                "summary": "Remove a repository",
                "parameters": [
                    {
                        "description": "Repository url",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/repo.RemovePayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.GlobalResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/app.GlobalResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.GlobalResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "app.GlobalResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "repo.ListData": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/repository.RepoData"
                    }
                }
            }
        },
        "repo.PrivateRepoDetails": {
            "type": "object",
            "properties": {
                "image_ref": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "repo.RemovePayload": {
            "type": "object",
            "properties": {
                "repo": {
                    "type": "string"
                }
            }
        },
        "repository.RepoData": {
            "type": "object",
            "properties": {
                "image_ref": {
                    "type": "string"
                },
                "reachable": {
                    "type": "boolean"
                },
                "url": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        }
    },
...
}